requirements.txt
numpy
scipy
sounddevice
openwakeword
silero-vad
python-dotenv
websockets>=12.0
